fix normals, LOD

+++ src/funcs.cpp
@@ -333,7 +333,7 @@ namespace funcs{
 	}
 
 
-	size_t flatten(size_t i, size_t j, size_t width){
+	int flatten(int i, int j, int width){
 		return (j + i * width);
 	}
